순환 의존을 막기 위해 token Provider를 따로 생성
