+++ src/main/java/com/example/smart_waiting/security/TokenProvider.java
@@ -1,0 +1,51 @@
+package com.example.smart_waiting.security;
+
+import com.example.smart_waiting.user.service.UserServiceImpl;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+import java.util.Date;
+import java.util.List;
+
+@Component
+@RequiredArgsConstructor
+public class TokenProvider {
+
+    private final UserServiceImpl userService;
+    private static final long TOKEN_EXPIRE_TIME = 1000*60*60*24; //1DAY
+    private static final String KEY_ROLES = "roles";
+    private final TokenUtil tokenUtil;
+
+    @Value("{spring.jwt.secret}")
+    private String secretKey;
+
+    public String generateToken(String username, List<String> roles){
+        Claims claims = Jwts.claims().setSubject(username);
+        claims.put(KEY_ROLES, roles);
+
+        var now = new Date();
+        var expiredDate = new Date(now.getTime() + TOKEN_EXPIRE_TIME);
+
+        return Jwts.builder()
+                .setClaims(claims)
+                .setIssuedAt(now)
+                .setExpiration(expiredDate)
+                .signWith(SignatureAlgorithm.HS512, secretKey)
+                .compact();
+    }
+
+
+    public Authentication getAuthentication(String jwt){
+
+        UserDetails userDetails = userService.loadUserByUsername(tokenUtil.getEmail(jwt));
+        return new UsernamePasswordAuthenticationToken(userDetails,"",userDetails.getAuthorities());
+    }
+
+}
