+++ src/main/java/com/example/smart_waiting/security/TokenUtil.java
@@ -1,58 +1,25 @@
 package com.example.smart_waiting.security;
 
-import com.example.smart_waiting.user.User;
-import com.example.smart_waiting.user.service.UserServiceImpl;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.Date;
-import java.util.List;
 
 @Component
-@RequiredArgsConstructor
 public class TokenUtil {
 
-    private static final long TOKEN_EXPIRE_TIME = 1000*60*60*24; //1DAY
-    private static final String KEY_ROLES = "roles";
     public static final String TOKEN_HEADER = "Authorization";
     public static final String TOKEN_PREFIX = "Bearer ";
-
-    private final UserServiceImpl userService;
 
     @Value("{spring.jwt.secret}")
     private String secretKey;
 
-    public String generateToken(String username, List<String> roles){
-        Claims claims = Jwts.claims().setSubject(username);
-        claims.put(KEY_ROLES, roles);
-
-        var now = new Date();
-        var expiredDate = new Date(now.getTime() + TOKEN_EXPIRE_TIME);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(now)
-                .setExpiration(expiredDate)
-                .signWith(SignatureAlgorithm.HS512, secretKey)
-                .compact();
-    }
-
-    public Authentication getAuthentication(String jwt){
-
-        UserDetails userDetails = userService.loadUserByUsername(getEmail(jwt));
-        return new UsernamePasswordAuthenticationToken(userDetails,"",userDetails.getAuthorities());
-    }
 
     public String getEmail(String token){
         return this.parseClaims(token).getSubject();
@@ -73,7 +40,7 @@ public class TokenUtil {
         }
     }
 
-    public String resolveTokenFromRequest(HttpServletRequest request){
+    public static String resolveTokenFromRequest(HttpServletRequest request){
         String token  = request.getHeader(TOKEN_HEADER);
 
         if(!ObjectUtils.isEmpty(token) && token.startsWith(TOKEN_PREFIX)){
